Rename dialogs.?pp
